requirements.txt
Flask
google-generativeai
google-api-python-client
langchain
langchain-community
pypdf
pymongo
python-dotenv
